pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
graphene-django = "3.2.0"
djangorestframework = "3.15.0"

[tool.ruff]
fix = true
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
